# Conflicts:
#	server/lobbyconnection.py
#	tests/unit_tests/test_lobbyconnection.py

+++ tests/data/test-data.sql
@@ -17,7 +17,7 @@ insert into ladder1v1_rating (id, mean, deviation, numGames, is_active) values
     (102, 1500, 500, 0, 1)
 ;
 
-delete from matchmaker_ban;
+delete from matchmaker_ban where id = 102 and userid = 102;
 insert into matchmaker_ban (id, userid) values (102, 102);
 
 insert into game_stats (id, startTime, gameType, gameMod, host, mapId, gameName, validity) values
@@ -38,3 +38,7 @@ insert into game_player_stats (gameId, playerId, AI, faction, color, team, place
     (41939, 1, 0, 0, 0, 2, 0, 1500, 500, NOW() + interval 4 minute),
     (41940, 1, 0, 0, 0, 2, 0, 1500, 500, NOW() + interval 5 minute),
     (41941, 1, 0, 0, 0, 2, 0, 1500, 500, NOW() + interval 6 minute);
+
+delete from friends_and_foes where user_id = 1 and subject_id = 2;
+insert into friends_and_foes (user_id, subject_id, status) values
+    (2, 1, 'FRIEND');
